--- Blood Donation Support/DTO/BloodCompatibilityResponseDto.cs
@@ -1,7 +1,0 @@
-namespace Blood_Donation_Support.DTO
-{
-    public class BloodCompatibilityResponseDto
-    {
-        public List<string> CompatibleBloodTypes { get; set; }
-    }
-}
