--- Blood Donation Support/DTO/BloodUnitStatusUpdateDTO.cs
@@ -1,7 +1,0 @@
-namespace Blood_Donation_Support.DTO
-{
-    public class BloodUnitStatusUpdateDTO
-    {
-        public string Status { get; set; }
-    }
-}
