--- Blood Donation Support/DTO/CreateReservationDto.cs
@@ -1,16 +1,0 @@
-using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace Blood_Donation_Support.DTO;
-
-public class CreateReservationDto
-{
-    [Required]
-    public int BloodUnitId { get; set; }
-
-    [Required]
-    public int TransfusionId { get; set; }
-
-    [Required]
-    public DateTime ExpireAt { get; set; }
-}
