+++ Blood Donation Support/DTO/HospitalDTO.cs
@@ -5,4 +5,5 @@ namespace Blood_Donation_Support.DTO
         public double Latitude { get; set; }
         public double Longitude { get; set; }
     }
+
 }
