+++ Blood Donation Support/DTO/ReservationDto.cs
@@ -1,8 +1,19 @@
-using System;
+using System.ComponentModel.DataAnnotations;
 
 namespace Blood_Donation_Support.DTO;
 
-public class ReservationDto
+public class CreateReservationDTO
+{
+    [Required]
+    public int BloodUnitId { get; set; }
+
+    [Required]
+    public int TransfusionId { get; set; }
+
+    [Required]
+    public DateTime ExpireAt { get; set; }
+}
+public class ReservationDTO
 {
     public int ReservationId { get; set; }
     public int BloodUnitId { get; set; }
@@ -11,3 +22,4 @@ public class ReservationDto
     public DateTime ExpireAt { get; set; }
     public string Status { get; set; } = string.Empty;
 }
+
