+++ Blood Donation Support/DTO/RoleDTO.cs
@@ -3,6 +3,6 @@ namespace Blood_Donation_Support.DTO
     public class RoleDTO
     {
         public int RoleId { get; set; }
-        public string Name { get; set; }
+        public string? Name { get; set; }
     }
 }
